refactor: move ORTHOGONAL_TASKS to constants.py
