+++ src/leaderboards/bootstrap_cis.py
@@ -16,7 +16,8 @@ from collections import defaultdict
 
 import numpy as np
 
-from .task_metadata import ORTHOGONAL_TASKS, task_category
+from .constants import ORTHOGONAL_TASKS
+from .task_metadata import task_category
 
 logger = logging.getLogger(__name__)
